update bn_convert script
